+++ src/index.ts
@@ -1,9 +1,4 @@
 export * from './distribution';
 export * from './heatmap';
 export * from './list';
-
-export * from './axis';
-export * from './box';
-export * from './force_directed_graph';
-export * from './kaplanmeier';
-export * from './table';
+export * from './base';
